requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
gunicorn
requests

